pyproject.toml
[project]
name = "realtime"
authors = [{name = "Haruto Nakayama", email = "nakayama.haruto@aist.go.jp"}]
readme = "README.md"
license = {file = "LICENSE"}
description = "Realtime-Processing-System-for-Crowd-Analysis"
version = "2026.04"
requires-python = ">=3.12.8,<3.13"

dependencies = [
  "matplotlib",
  "lightning",
  "numpy",
  "pyyaml",
  "torch",
  "torchaudio",
  "torchmetrics",
  "torchvision",
  "albumentations",
  "hydra-core",
  "omegaconf",
  "tqdm",
  "pydantic",
  "timm",
  "pandas",
  "websockets",
  "asyncio",
  "onvif-zeep",
]


# -------------------------
# pixi: conda 側（OS/ネイティブ/ツール）
# -------------------------
[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.dependencies]
python = ">=3.12.8,<3.13"
ffmpeg = "*"

# ここは conda の方が安定しやすい（ネイティブ依存が絡む）
scipy = "*"
cython = "*"
opencv = "*"        # opencv-python (pip) ではなく conda の opencv を推奨
lap = "*"
sympy = "*"

# -------------------------
# pixi: PyPI 側（pipで入れるもの）
# -------------------------
[tool.pixi.pypi-dependencies]
nvidia-dali-cuda120 = { version = "*", index = "https://pypi.nvidia.com" }